feat: created a web component for the bridge deposit and withdraw component
