+++ cli/lib/index.mjs
@@ -1,9 +1,8 @@
 /* ignore unused exports */
 import Nf3 from './nf3.mjs';
-import getAbi from './abi.mjs';
 import * as Constants from './constants.mjs';
 import * as Tokens from './tokens.mjs';
 import * as Units from './units.mjs';
 import * as Environment from './environment.mjs';
 
-export { Nf3, Constants, Tokens, Units, Environment, getAbi };
+export { Nf3, Constants, Tokens, Units, Environment };
