+++ nightfall-deployer/entrypoint.sh
@@ -2,7 +2,7 @@
 set -o errexit
 set -o pipefail
 
-if [ -z "${USE_INFURA}" ] && [ -z "${ETH_PRIVATE_KEY}" ] && [ -z "${USE_ROPSTEN_NODE}" ]; then
+if [ -z "${USE_INFURA}" ] && [ -z "${ETH_PRIVATE_KEY}" ]; then
   # wait until there's a blockchain client up
   while ! nc -z ${BLOCKCHAIN_WS_HOST} ${BLOCKCHAIN_PORT}; do sleep 3; done
 fi
