+++ nightfall-deployer/src/circuit-setup.mjs
@@ -134,7 +134,7 @@ async function setupCircuits() {
 
       // when deploying on infura - do serial tx execution to avoid nonce issue
       // when using a private key, we shouldn't assume an unlocked account and we sign the transaction directly
-      if (config.USE_INFURA || config.ETH_PRIVATE_KEY || config.USE_ROPSTEN_NODE) {
+      if (config.USE_INFURA || config.ETH_PRIVATE_KEY) {
         await Web3.submitRawTransaction(await tx.encodeABI(), keyRegistryAddress);
       } else await tx.send();
     } catch (err) {
