+++ package.json
@@ -40,13 +40,15 @@
     "config": "^3.3.1",
     "eslint": "^7.25.0",
     "eslint-config-codfish": "^10.0.0",
+    "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.26.1",
     "fast-check": "^2.18.0",
     "find-unused-exports": "^2.0.0",
     "general-number": "^1.0.1",
     "husky": "^6.0.0",
-    "jsdoc": "^3.6.10",
+    "jsdoc": "^3.6.7",
     "mocha": "^8.3.2",
     "prettier": "^2.2.1",
     "prettier-plugin-solidity": "^1.0.0-beta.10",
