+++ wallet/.env
@@ -1,5 +1,5 @@
 SKIP_PREFLIGHT_CHECK=true
-OPTIMIST_API_URL=http://ohio.nightfall3.com:80
+OPTIMIST_API_URL=https://vqxy02tr5e.execute-api.us-east-2.amazonaws.com/Dev
 LOCAL_API_URL=http://localhost:8081
 LOCAL_OPTIMIST=true
 OPTIMIST_WS_URL=wss://af3ys26lpe.execute-api.us-east-2.amazonaws.com/dev
