+++ wallet/package.json
@@ -29,10 +29,13 @@
     "react-responsive": "^9.0.0-beta.6",
     "react-router-dom": "5.3.0",
     "react-scripts": "4.0.3",
+    "react-to-webcomponent": "^1.5.1",
     "redux": "4.1.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
-    "sass": "^1.49.0",
+    "safe-buffer": "^5.2.1",
+    "sass": "^1.49.7",
+    "typescript": "^4.5.5",
     "wasm-loader": "^1.3.0",
     "web-vitals": "^2.1.0",
     "web3": "^1.6.0",
