+++ wallet/src/DTO/IUser.tsx
@@ -1,0 +1,10 @@
+interface IUser {
+  active: false;
+  zkpKeys: {
+    pkd: string;
+    nsk: string;
+    ask: string;
+  };
+}
+
+export default IUser;
