+++ wallet/src/custom.d.ts
@@ -1,0 +1,4 @@
+declare module "*.svg";
+declare module "*.png";
+declare module '*.scss';
+declare module '*.sass';
