+++ wallet/src/hooks/User/index.jsx
@@ -11,20 +11,25 @@ import { getMaxBlock } from '../../nightfall-browser/services/database';
 
 const { eventWsUrl } = global.config;
 
-export const reducer = (state, action) => {
-  switch (action.type) {
-    case 'toggle_button':
-      return {
-        ...state,
-        active: !state.active,
-      };
-    default:
-      return state;
-  }
-};
+// export const reducer = (state, action) => {
+//   switch (action.type) {
+//     case 'toggle_button':
+//       return {
+//         ...state,
+//         active: !state.active,
+//       };
+//     default:
+//       return state;
+//   }
+// };
 
 export const initialState = {
   active: false,
+  zkpKeys: {
+    pkd: '',
+    nsk: '',
+    ask: '',
+  },
 };
 
 export const UserContext = React.createContext({
